Fix the bug that the state of replica is always in decommission (#1997) (#2143)
To remove the redundant replica, proceed as follows.
1. Select the replica to be dropped according to the following priority
   1). replica on the backend, which was dropped.
   2). replica state is bad
   3). replica is unavailable
   4). replica' state is under clone or decommission.
   5). replica with failed version.
   6). replica with a lower version.
   7). replica on the same host.
   8). replica as the balance source.
   9). replica stored on high load backend.
2. Set the replica state to decommission, find the current largest txn number and return the TabletSchedCtx to the scheduling queue.
3. All the txn smaller than the found txn number is finished. FE can drop this replica.

But the second and third steps may be separated by a long time. Another replica sees the problem, like the version incomplete or backend unavailable during this period. The scheduler may go through other repair processes or select another replica to decommission if this happens. The replica state will stay as decommissioned and not be scheduled.

To fix this problem, reset the decommissioned replica state when
1. Scheduler chooses another replica to decommission.
2. TabletSchedCtx is finished.