Fix query error in global dictionary when replicas are inconsistent
In some cases where replicas are inconsistent, the segment may
not have a dictionary encoding page, but at this time,
the FE holds the global dictionary, which will cause the query to fail.

We need to make some adjustments
Add a stage, when a column of the final chunk is a string
and the column needs to use the global dictionary,
we convert the string of this column to the global dictionary encoding.